Move RemainingCharacters to components
