+++ front/app/containers/Admin/messaging/texting/NewCampaign/index.tsx
@@ -7,7 +7,7 @@ import { Label, Box, IconTooltip } from '@citizenlab/cl2-component-library';
 import { Section, SectionField } from 'components/admin/Section';
 import HelmetIntl from 'components/HelmetIntl';
 import TextingHeader from '../components/TextingHeader';
-import RemainingCharacters from '../NewSMSMessage/RemainingCharacters';
+import RemainingCharacters from '../components/RemainingCharacters';
 
 // services
 import { addTextingCampaign } from 'services/textingCampaigns';
